Upgrade to .NET 10.0 and refactor LINQ evaluation
Upgraded all projects to target .NET 10.0, enabling the latest C# language features (`LangVersion: latest`). Refactored `SqlExpressionVisitor` to replace `Expression.Compile().DynamicInvoke()` with safer and more efficient evaluation methods (`EvaluateValue`, `EvaluateEnumerable`, `ResolveCollection`). Improved handling of LINQ `Contains` and string methods, supporting more patterns and ensuring compatibility with .NET 10.0.

Normalized enum values to their underlying integer types for SQL parameterization, ensuring database compatibility. Enhanced test coverage for LINQ scenarios and added detailed logging for SQL parameter inspection. Updated dependencies to their latest versions and consolidated target frameworks where applicable.

Reformatted project files for consistency, removed redundant code, and introduced helper methods to improve maintainability. These changes ensure compatibility with .NET 10.0, improve performance, and enhance code readability.

+++ ObjectInfo/SpecificationGenerator.Tests.Dapper/Tests/QueryGenerationTests.cs
@@ -175,8 +175,9 @@ namespace ObjectInfo.Deepdive.SpecificationGenerator.Tests.Dapper.Tests
             sql.Should().Contain("CustomerType IN (@p0, @p1)");
 
             var parameters = spec.GetParameters();
-            parameters["@p0"].Should().Be(CustomerType.Premium);
-            parameters["@p1"].Should().Be(CustomerType.VIP);
+            // Enum values are normalized to their underlying integer type for SQLite compatibility
+            parameters["@p0"].Should().Be((int)CustomerType.Premium);
+            parameters["@p1"].Should().Be((int)CustomerType.VIP);
         }
 
         // Note: Paging and ordering functionalities are not implemented in the current code.
